Rename `__e2e__` to `e2e` and follow file naming convention
